+++ docs/.vitepress/config.ts
@@ -2,7 +2,6 @@ import Unocss from 'unocss/vite'
 import svgLoader from 'vite-svg-loader'
 import { defineConfig } from 'vitepress'
 import { resolve } from 'pathe'
-import Components from 'unplugin-vue-components/vite'
 
 export default defineConfig({
   title: 'TresJS',
@@ -96,13 +95,7 @@ export default defineConfig({
     ],
   },
   vite: {
-    plugins: [
-      svgLoader(),
-      Unocss(),
-      Components({
-        dirs: ['./theme/components'],
-      }),
-    ],
+    plugins: [svgLoader(), Unocss()],
     resolve: {
       alias: {
         '/@': resolve(__dirname, '../../src'),