+++ src/core/nodeOps.ts
@@ -1,21 +1,15 @@
-import { BufferAttribute, BufferGeometry, Material, Mesh, Object3D } from 'three'
-import { useCamera, useRaycaster, useRenderLoop, useLogger } from '/@/composables'
+import { BufferAttribute } from 'three'
+import { useCamera } from '/@/composables'
 import { RendererOptions } from 'vue'
 import { catalogue } from './catalogue'
-import { isFunction, useEventListener } from '@vueuse/core'
-import { TresEvent, TresObject } from '../types'
+import { isFunction } from '@vueuse/core'
+import { TresObject } from '../types'
 import { isHTMLTag, kebabToCamel } from '../utils'
 
-const { logWarning } = useLogger()
+import type { Object3D, Material, BufferGeometry } from 'three'
 
-function hasEvents(obj: any) {
-  for (const prop in obj) {
-    if (prop.indexOf('on') === 0) {
-      return true
-    }
-  }
-  return false
-}
+const onRE = /^on[^a-z]/
+export const isOn = (key: string) => onRE.test(key)
 
 function noop(fn: string): any {
   fn
@@ -43,12 +37,11 @@ export const nodeOps: RendererOptions<TresObject, TresObject> = {
     }
 
     if (instance.isCamera) {
-      // Let users know that camera is in the center of the scene
-      if (!props?.position || props?.position.every((v: number) => v == 0)) {
-        logWarning(
-          // eslint-disable-next-line max-len
-          'Camera is positioned at the center of the scene [0,0,0], if this is not intentional try setting a position if your scene seems empty 🤗',
-        )
+      if (!props?.position) {
+        instance.position.set(3, 3, 3)
+      }
+      if (!props?.lookAt) {
+        instance.lookAt(0, 0, 0)
       }
       const { pushCamera } = useCamera()
       pushCamera(instance)
@@ -68,6 +61,8 @@ export const nodeOps: RendererOptions<TresObject, TresObject> = {
       if (props?.geometry?.isBufferGeometry) (instance as Object3D).userData[GEOMETRY_VIA_PROP] = true
     }
 
+    instance.events = {}
+
     return instance
   },
   insert(child, parent, anchor) {
@@ -85,43 +80,6 @@ export const nodeOps: RendererOptions<TresObject, TresObject> = {
       if (parent) {
         parent[child.attach] = child
       }
-    }
-
-    const { onLoop } = useRenderLoop()
-
-    // RayCasting
-    let prevInstance: TresEvent | null = null
-    let currentInstance: TresEvent | null = null
-
-    if (child && child instanceof Mesh && hasEvents(child)) {
-      const { raycaster } = useRaycaster()
-      onLoop(() => {
-        if (parent?.children && child && raycaster) {
-          const intersects = raycaster.value.intersectObjects(parent.children)
-
-          if (intersects.length > 0 && intersects[0].object.uuid === child.uuid) {
-            currentInstance = intersects[0]
-
-            if (prevInstance === null || prevInstance.object.uuid !== currentInstance?.object.uuid) {
-              child.onPointerEnter?.(currentInstance)
-            }
-
-            child.onPointerMove?.(currentInstance)
-          } else {
-            currentInstance = null
-            if (prevInstance !== null) {
-              child.onPointerLeave?.(prevInstance)
-            }
-          }
-
-          prevInstance = currentInstance
-        }
-      })
-
-      useEventListener(window, 'click', () => {
-        if (currentInstance === null) return
-        child.onClick?.(currentInstance)
-      })
     }
   },
   remove(node) {
@@ -176,7 +134,9 @@ export const nodeOps: RendererOptions<TresObject, TresObject> = {
 
         if (!target?.set) root = chain.reduce((acc, key) => acc[kebabToCamel(key)], root)
       }
-
+      if (isOn(key)) {
+        node.events[key] = nextValue
+      }
       let value = nextValue
       if (value === '') value = true
       // Set prop, prefer atomic methods if applicable